requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
google-api-core==1.23.0
google-auth==1.24.0
google-cloud-language==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.15
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
spotipy==2.16.1
SQLAlchemy==1.3.22
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
